Refactor simulation interface into modular league coordinator
